Woah! An Update!
In this update, I completely overhauled the plugin. I scrapped 90% of my old ideas and work. That of course means there is barely anything here right now, the purpose of pushing this update to GitHub is honestly purely for my own benefit. I'm pushing out progress! This sort of commit, containing varying amounts of unfinished and broken content is going to be normal until the first real release. You can see some details of my plans in the updated Readme.